Resolved #257: added pass-through for sending data to Access
Blip: a "pass through" function was installed for scriptwriters. To
create a message on the Access database without creating a corresponding
pop-up, simply use `script_end_procedure("~PT: [message goes here]")`.
The contents of the message will be passed through to the database
without showing up as a pop-up.

@courtrightd, @C-Love, @IlseFerris, @gays, @CDPotter, and
@RobertFewins-Kalb may find this interesting, as I know your agencies
use the database (or are in the process of doing so).